start to restyle the login screen for CDS styles
