+++ localization/react-intl/src/app/components/UserTosForm.json
@@ -5,7 +5,7 @@
     "defaultMessage": "Terms of Service"
   },
   {
-    "id": "userTos.ppLink",
+    "id": "userTos.ppLink\"",
     "description": "Link text to take the user to the privacy policy",
     "defaultMessage": "Privacy Policy"
   },
