+++ src/app/components/login/login.module.css
@@ -1,0 +1,21 @@
+.login-wrapper {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  width: 100%;
+
+  .login-container {
+    border: solid 2px var(--validationMain);
+    margin: 32px auto 16px;
+    max-width: 680px;
+    padding: 0 16px;
+    width: 100%;
+
+    .login-form {
+      background-color: var(--otherWhite);
+      border-radius: 16px;
+      margin: 16px auto;
+      padding: 16px;
+    }
+  }
+}
